requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.1.1
django-cors-headers==3.5.0
django-rest-framework==0.1.0
djangorestframework==3.11.1
flake8==3.8.3
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1593446406207/work
mccabe==0.6.1
psycopg2 @ file:///tmp/build/80754af9/psycopg2_1594305108220/work
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
